+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Trim terms in JAutoCompletion [#470](https://github.com/nbbrd/java-desktop-util/issues/470)
+
 ### Changed
 
 - Modernize use of NIO API